# Speedup compilation