--- a/.gitignore
+++ b/.gitignore
@@ -4,3 +4,4 @@ CMakeFiles/
 build
 compile_commands.json
 .DS_Store
+.cache