--- a/Common/Core/dummy.cxx
+++ b/Common/Core/dummy.cxx
@@ -1,0 +1,5 @@
+#include "Framework/WorkflowSpec.h"
+using namespace o2::framework;
+WorkflowSpec defineDataProcessing(o2::framework::ConfigContext const&){
+  return WorkflowSpec{};
+}