@alibuild 2022-03-11
Error while checking build/O2Physics/o2 for f05ef01a9cddd13d0f9df3ac39d588c3d40196ce at 2022-03-11 16:14:
```
## sw/BUILD/O2Physics-latest/log
cc1plus: error: /sw/BUILD/f9108e7cdc516d4ae2b6d05d1904bcb911cf4974/O2Physics/Common/Core/CMakeFiles/O2Physicslib-PrecompiledO2Framework.dir/cmake_pch.hxx.gch: not used because `O2Physicslib_PrecompiledO2Framework_EXPORTS' not defined [-Werror=invalid-pch]
cc1plus: error: /sw/BUILD/f9108e7cdc516d4ae2b6d05d1904bcb911cf4974/O2Physics/Common/Core/CMakeFiles/O2Physicslib-PrecompiledO2Framework.dir/cmake_pch.hxx.gch: not used because `O2Physicslib_PrecompiledO2Framework_EXPORTS' not defined [-Werror=invalid-pch]
cc1plus: error: /sw/BUILD/f9108e7cdc516d4ae2b6d05d1904bcb911cf4974/O2Physics/Common/Core/CMakeFiles/O2Physicslib-PrecompiledO2Framework.dir/cmake_pch.hxx.gch: not used because `O2Physicslib_PrecompiledO2Framework_EXPORTS' not defined [-Werror=invalid-pch]
cc1plus: error: /sw/BUILD/f9108e7cdc516d4ae2b6d05d1904bcb911cf4974/O2Physics/Common/Core/CMakeFiles/O2Physicslib-PrecompiledO2Framework.dir/cmake_pch.hxx.gch: not used because `O2Physicslib_PrecompiledO2Framework_EXPORTS' not defined [-Werror=invalid-pch]
cc1plus: error: /sw/BUILD/f9108e7cdc516d4ae2b6d05d1904bcb911cf4974/O2Physics/Common/Core/CMakeFiles/O2Physicslib-PrecompiledO2Framework.dir/cmake_pch.hxx.gch: not used because `O2Physicslib_PrecompiledO2Framework_EXPORTS' not defined [-Werror=invalid-pch]
cc1plus: error: /sw/BUILD/f9108e7cdc516d4ae2b6d05d1904bcb911cf4974/O2Physics/Common/Core/CMakeFiles/O2Physicslib-PrecompiledO2Framework.dir/cmake_pch.hxx.gch: not used because `O2Physicslib_PrecompiledO2Framework_EXPORTS' not defined [-Werror=invalid-pch]
ninja: build stopped: subcommand failed.
```
Full log [here](https://ali-ci.cern.ch/alice-build-logs/AliceO2Group/O2Physics/608/f05ef01a9cddd13d0f9df3ac39d588c3d40196ce/build_O2Physics_o2/pretty.html).